Add files via upload
The use of machine learning techniques for
assessing the potential of organizational
resilience

Organizational resilience (OR) increases when the company has the ability
to anticipate, plan, make decisions, and react quickly to changes and
disruptions. Thus the company should focus on the creation and implementation
of proactive and innovative solutions. Proactive processing of
information requires modern technological solutions and new techniques
used. The main focus of this study is to propose the best technique of Machine Learning (ML) in the context of accuracy for predicting the
attributes of the organizational resilience potential. Based on the calculations,
the research includes estimating them through the applications
of regression and machine learning methods. The dataset is obtained
from the results of the our survey based on the questionnaire consisting
of 48 items mainly established on OR attributes formed on ISO
22316:2017 standard. Based on the outcomes of the study, it can be stated
that the optimal technique in the context of accuracy for predicting the
attributes of the organizational resilience potential is ensemble methods.
The k-nearest neighbor (KNN) filtering-based data pre-processing technique
for stacked ensemble classifier is used. The stacking is achieved
with three base classifiers namely Random Forest (RF), Naive Bayes
(NB), and Support Vector Machine (SVM). The chosen technique
should be implemented in an organization systemically according to
the circle of innovation, and should support the quality of managerial
decision-making process by increasing the accuracy of organizational
resilience potential prediction, and indication of the importance of
attributes and factors affecting the potential for organizational resilience.

To cite this project
Ewertowski, T., Guldogus, B.C., Kuter, S., Akyuz, S. Weber, G.W., Wrzesinska, J.S., Racek, E. The use of machine learning techniques for assessing the potential of organizational resilience. 2023.